Updated to 1.20.1 / Added Brick Forge (without Functionality)
Updated the mod to 1.20.1, which was successful. Implemented a custom block; however, the block doesn't actually do anything yet and if the player is facing more than 45 degrees down or up when placing the brick forge, the game will just crash. Not sure if that would be a client or server side issue, but it would likely be server side.